requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.0
celery==5.2.7
certifi==2020.11.8
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.1
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-encrypted-fields==1.1.2
django-fernet-fields==0.6
django-redis==5.2.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.5
Faker==15.3.1
filelock==3.8.0
Fiona==1.9.4.post1
Flask==2.3.3
folium==0.14.0
fontawesomefree==6.3.0
fonttools==4.42.1
frozenlist==1.3.3
fuzzywuzzy==0.18.0
geopandas==0.13.2
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.0
Levenshtein==0.21.1
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
packaging==21.3
panda==0.3.1
pandas==2.0.3
Pillow==9.4.0
pinecone-client==2.2.1
platformdirs==2.5.2
playsound==1.2.2
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.18.0
PyJWT==2.7.0
pyparsing==3.0.9
pyproj==3.6.0
pytest==7.1.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-keyczar==0.716
python-Levenshtein==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.2.0
readability-lxml==0.8.1
redis==4.5.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.1
shapely==2.0.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4
sqlparse==0.4.4
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.7
yarl==1.8.2
zipp==3.16.2
